Rename blueprint to fake_audio_detection
