+++ .github/ISSUE_TEMPLATE/feature_request.yaml
@@ -2,7 +2,7 @@ name: "\U0001F680 Feature Request"
 description: Suggest an idea for this project
 title: "[FEATURE]: "
 projects:
-  - "Project-name"
+  - "fake-audio-detection"
 labels:
   - enhancement
 body:
@@ -10,7 +10,7 @@ body:
     attributes:
       value: |
         Thanks for taking the time to fill out this form!
-        Please make sure you have searched for a similar [issue](https://github.com/mozilla-ai/project-name/issues) before submitting a new one.
+        Please make sure you have searched for a similar [issue](https://github.com/mozilla-ai/fake-audio-detection/issues) before submitting a new one.
 
   - type: textarea
     id: motivation
@@ -40,7 +40,7 @@ body:
       label: Contribution
       description: |
         Is there any way that you could help, e.g. by submitting a PR?
-        Make sure to read the [contribution guidelines](https://github.com/mozilla-ai/project-name/blob/main/CONTRIBUTING.md).
+        Make sure to read the [contribution guidelines](https://github.com/mozilla-ai/fake-audio-detection/blob/main/CONTRIBUTING.md).
       placeholder: |
         I could help by...
     validations:
