+++ demo/app.py
@@ -9,7 +9,7 @@ import streamlit as st
 parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(parent_dir)
 
-from src.blueprint.model import predict_audio_blocks
+from src.fake_audio_detection.model import predict_audio_blocks
 
 st.title("🔎 DeepVoice Detection")
