+++ mkdocs.yml
@@ -1,6 +1,6 @@
 site_name: Guidance Docs
-repo_url: https://github.com/mozilla-ai/blueprint-template
-repo_name: blueprint-template
+repo_url: https://github.com/mozilla-ai/fake-audio-detection/
+repo_name: fake_audio_detection
 
 nav:
   - Home: index.md
