+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=48", "setuptools_scm[toml]>=6.3.1"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "blueprint"
+name = "fake-audio-detection"
 readme = "README.md"
 license = {text = "Apache-2.0"}
 requires-python = ">=3.10"
@@ -33,9 +33,9 @@ tests = [
 ]
 
 [project.urls]
-Documentation = "https://mozilla-ai.github.io/Blueprint-template/"
-Issues = "https://github.com/mozilla-ai/Blueprint-template/issues"
-Source = "https://github.com/mozilla-ai/Blueprint-template"
+Documentation = "https://mozilla-ai.github.io/fake-audio-detection/"
+Issues = "https://github.com/mozilla-ai/fake-audio-detection/issues"
+Source = "https://github.com/mozilla-ai/fake-audio-detection"
 
 [tool.setuptools.packages.find]
 exclude = ["tests", "tests.*"]
